% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{launch_data_viewer}
\alias{launch_data_viewer}
\title{View a data frame in the Data Viewer application}
\usage{
launch_data_viewer(
  data,
  name = NULL,
  use_parquet = FALSE,
  port = 3000,
  timeout = 3,
  app_path = "/Applications/data_viewer.app/Contents/MacOS/data_viewer"
)
}
\arguments{
\item{data}{A data frame to view.}

\item{name}{An optional name for the data frame in the viewer.}

\item{use_parquet}{A logical value indicating whether to use the Parquet format. Defaults to FALSE.}

\item{port}{The port number for the Data Viewer application. Defaults to 3000.}

\item{timeout}{The timeout in seconds for the application to start. Defaults to 3.}

\item{app_path}{The path to the Data Viewer application executable.}
}
\description{
This function launches the Data Viewer application and displays the given data frame.
}
