% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dataviewers.R
\name{list_dataviewers}
\alias{list_dataviewers}
\title{List Active Background Dataviewer Processes}
\usage{
list_dataviewers()
}
\description{
List Active Background Dataviewer Processes
}
\seealso{
For more information, please refer to the \href{https://madhankumarnagaraji.github.io/dataviewR/articles/Exporting-and-Reproducibility.html#to-know-the-session-information}{documentation}.
}
