% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_dataviewer.R
\name{stop_dataviewer}
\alias{stop_dataviewer}
\title{Stop a Specific or Most Recent Dataviewer Process}
\usage{
stop_dataviewer(id = NULL)
}
\arguments{
\item{id}{Character string specifying the process ID to stop.
If NULL, stops the most recent background dataviewer.}
}
\description{
Stop a Specific or Most Recent Dataviewer Process
}
\seealso{
For more information, please
refer to the \href{https://madhankumarnagaraji.github.io/dataviewR/articles/Exporting-and-Reproducibility.html#to-stop-dataviewr-sessions}{documentation}.
}
