% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviewer.R
\name{stop_dataviewer}
\alias{stop_dataviewer}
\title{Stop a Background Dataviewer Process}
\usage{
stop_dataviewer(id = NULL)
}
\arguments{
\item{id}{Character string specifying the process ID to stop. If NULL, stops the most recent background dataviewer.}
}
\description{
Stop a Background Dataviewer Process
}
